Version 0.1.7 ------------- prototype html engine

+++ thor/src/jvmMain/kotlin/io/github/remmerw/thor/cobra/util/io/IORoutines.kt
@@ -161,8 +161,8 @@ object IORoutines {
                 sb = StringBuffer()
             }
             when (b) {
-                '\n'.code.toByte() -> break@OUTER
-                '\r'.code.toByte() -> {}
+                '\n'.code -> break@OUTER
+                '\r'.code -> {}
                 else -> sb.append(b.toChar())
             }
         }
